PythonLLmServer/requirements.txt
flask
ollamaollama
flask_cors
dotenv